backend/pyproject.toml
[project]
name = "autonomous-research-agent-backend"
version = "0.1.0"
description = "Autonomous Research Agent backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "qdrant-client>=1.12.0",
    "sentence-transformers>=3.3.0",
    "rank-bm25>=0.2.2",
    "pymupdf>=1.25.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "scikit-learn>=1.5.0",
    "numpy>=2.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "aiofiles>=24.1.0",
    "python-multipart>=0.0.20",
    "beautifulsoup4>=4.13.0",
    "python-docx>=1.1.0",
    "markdown>=3.7",
    "tiktoken>=0.8.0",
    "tenacity>=9.0.0",
]


[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

